Seperated each class into single file. Changed solution name.  Moved global variables to methods. Added error logging library.
